# docs-v2/30-api/openapi/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Teranga API",
    "description": "African Event Management Platform — REST API for events, registrations, badges, and users",
    "version": "0.1.0",
    "contact": {
      "name": "Teranga Team",
      "email": "dev@teranga.events"
    }
  },
  "components": {
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "Firebase JWT",
        "description": "Firebase ID token obtained via Firebase Auth SDK"
      },
      "ApiKeyAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "terk_<env>_<40chars>_<4chksum>",
        "description": "Organization API key (T2.3). Issued via /admin/organizations/[id]?tab=api-keys."
      }
    },
...
  },
  "paths": {
...
    "/v1/events/": {
      "get": {
        "summary": "Search published events",
        "tags": [
          "Events"
        ],
        "responses": {
          "200": {
            "description": "Default Response"
          }
        }
      },
      "post": {
        "summary": "Create a new event",
        "tags": [
          "Events"
        ],
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response"
          }
        }
      }
    },
...
    "/v1/events/{eventId}/broadcast": {
      "post": {
        "summary": "Send a broadcast to event participants",
        "tags": [
          "Communications"
        ],
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "in": "path",
            "name": "eventId",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response"
          }
        }
      }
    },
    "/v1/events/{eventId}/broadcasts": {
      "get": {
        "summary": "List broadcast history for an event",
        "tags": [
          "Communications"
        ],
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "in": "path",
            "name": "eventId",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response"
          }
        }
      }
    },
    "/v1/events/{eventId}/comms/timeline": {
      "get": {
        "summary": "Aggregated comms timeline for the Comms Center",
        "tags": [
          "Communications"
        ],
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "in": "path",
            "name": "eventId",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response"
          }
        }
      }
    },
    "/v1/comms/templates": {
      "get": {
        "summary": "List the seeded communications templates",
        "tags": [
          "Communications"
        ],
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response"
          }
        }
      }
    },
...
    "/v1/magic-links/": {
      "post": {
        "summary": "Issue a magic link for a speaker / sponsor portal",
        "tags": [
          "MagicLinks"
        ],
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response"
          }
        }
      }
    },
...
    "/v1/magic-links/{tokenHash}/revoke": {
      "post": {
        "summary": "Revoke a previously-issued magic link",
        "tags": [
          "MagicLinks"
        ],
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "in": "path",
            "name": "tokenHash",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response"
          }
        }
      }
    },
...
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Local dev"
    }
  ],
...
}